+++ lib/cache/redis_test.go
@@ -4,7 +4,6 @@ package cache
 
 import (
 	"fmt"
-	"github.com/stretchr/testify/assert"
 	"testing"
 	"time"
 
@@ -12,6 +11,7 @@ import (
 	"github.com/Jeffail/benthos/v3/lib/metrics"
 	"github.com/Jeffail/benthos/v3/lib/types"
 	"github.com/ory/dockertest/v3"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestRedisIntegration(t *testing.T) {